0.1v  Version is Uploaded
This is a Shopping list project in c++ which lists the product along with their code and price. We can see the price of the product along with their respective price, can delete the product. And we can keep doing the things in it until we want to Quit on our own.